move isobased functionality from routingalg.py to isobased.py
