+++ Isochrone/algorithms/isobased.py
@@ -12,6 +12,7 @@ from scipy.stats import binned_statistic
 import utils.graphics as graphics
 import utils.formatting as form
 from ship.ship import Boat
+from ship.shipparams import ShipParams
 from algorithms.routingalg import RoutingAlg
 from routeparams import RouteParams
 from weather import WeatherCond
@@ -19,15 +20,199 @@ from weather import WeatherCond
 logger = logging.getLogger('WRT.Pruning')
 
 class IsoBased(RoutingAlg):
+    is_last_step: bool
+    is_pos_constraint_step: bool
+
+    start_temp: tuple
+    finish_temp: tuple
+    gcr_azi_temp: tuple
+
+    '''
+           All variables that are named *_per_step constitute (M,N) arrays, whereby N corresponds to the number of variants (plus 1) and
+           M corresponds to the number of routing steps.
+
+           At the start of each routing step 'count', the element(s) at the position 'count' of the following arrays correspond to
+           properties of the point of departure of the respective routing step. This means that for 'count = 0' the elements of
+           lats_per_step and lons_per_step correspond to the coordinates of the departure point of the whole route. The first 
+           elements of the attributes 
+               - azimuth_per_step
+               - dist_per_step
+               - speed_per_step
+           are 0 to satisfy this definition.
+       '''
+
+    lats_per_step: np.ndarray  # lats: (M,N) array, N=headings+1, M=steps (M decreasing)    #
+    lons_per_step: np.ndarray  # longs: (M,N) array, N=headings+1, M=steps
+    azimuth_per_step: np.ndarray  # heading
+    dist_per_step: np.ndarray  # geodesic distance traveled per time stamp:
+    shipparams_per_step: ShipParams
+    starttime_per_step: np.ndarray
+
+    current_azimuth: np.ndarray  # current azimuth
+    current_variant: np.ndarray  # current variant
+
+    # the lenght of the following arrays depends on the number of variants (variant segments)
+    full_dist_traveled: np.ndarray  # full geodesic distance since start for all variants
+    full_time_traveled: np.ndarray  # time elapsed since start for all variants
+    full_fuel_consumed: np.ndarray
+    time: np.ndarray  # current datetime for all variants
+
+    variant_segments: int  # number of variant segments in the range of -180° to 180°
+    variant_increments_deg: int
+    expected_speed_kts: int
+    prune_sector_deg_half: int  # angular range of azimuth that is considered for pruning (only one half)
+    prune_segments: int  # number of azimuth bins that are used for pruning
 
     def __init__(self, start, finish, time, figurepath=""):
-        RoutingAlg.__init__(self, start, finish, time, figurepath)
+        super().__init__(self, start, finish, figurepath)
+
+        self.lats_per_step = np.array([[start[0]]])
+        self.lons_per_step = np.array([[start[1]]])
+        self.azimuth_per_step = np.array([[None]])
+        self.dist_per_step = np.array([[0]])
+        sp = ShipParams.set_default_array()
+        self.shipparams_per_step = sp
+        self.starttime_per_step = np.array([[time]])
+
+        self.time = np.array([time])
+        self.full_time_traveled = np.array([0])
+        self.full_fuel_consumed = np.array([0])
+        self.full_dist_traveled = np.array([0])
+
         self.current_variant=self.current_azimuth
         self.is_last_step = False
         self.is_pos_constraint_step = False
 
     def print_init(self):
         RoutingAlg.print_init(self)
+
+    def print_shape(self):
+        print('PRINTING SHAPE')
+        print('per-step variables:')
+        print('     lats_per_step = ', self.lats_per_step.shape)
+        print('     lons_per_step = ', self.lons_per_step.shape)
+        print('     azimuths = ', self.azimuth_per_step.shape)
+        print('     dist_per_step = ', self.dist_per_step.shape)
+
+        self.shipparams_per_step.print_shape()
+
+        print('per-variant variables:')
+        print('     time =', self.time.shape)
+        print('     full_dist_traveled = ', self.full_dist_traveled.shape)
+        print('     full_time_traveled = ', self.full_time_traveled.shape)
+        print('     full_fuel_consumed = ', self.full_fuel_consumed.shape)
+
+    def current_position(self):
+        print('CURRENT POSITION')
+        print('lats = ', self.current_lats)
+        print('lons = ', self.current_lons)
+        print('azimuth = ', self.current_azimuth)
+        print('full_time_traveled = ', self.full_time_traveled)
+
+    def get_current_lats(self):
+        pass
+
+    def get_current_lons(self):
+        pass
+
+    def get_current_speed(self):
+        pass
+
+    def define_variants(self):
+        # branch out for multiple headings
+        nof_input_routes = self.lats_per_step.shape[1]
+
+        new_finish_one = np.repeat(self.finish_temp[0], nof_input_routes)
+        new_finish_two = np.repeat(self.finish_temp[1], nof_input_routes)
+
+        new_azi = geod.inverse(
+            self.lats_per_step[0],
+            self.lons_per_step[0],
+            new_finish_one,
+            new_finish_two
+        )
+
+        self.lats_per_step = np.repeat(self.lats_per_step, self.variant_segments + 1, axis=1)
+        self.lons_per_step = np.repeat(self.lons_per_step, self.variant_segments + 1, axis=1)
+        self.dist_per_step = np.repeat(self.dist_per_step, self.variant_segments + 1, axis=1)
+        self.azimuth_per_step = np.repeat(self.azimuth_per_step, self.variant_segments + 1, axis=1)
+        self.starttime_per_step = np.repeat(self.starttime_per_step, self.variant_segments + 1, axis=1)
+
+        self.shipparams_per_step.define_variants(self.variant_segments)
+
+        self.full_time_traveled = np.repeat(self.full_time_traveled, self.variant_segments + 1, axis=0)
+        self.full_fuel_consumed = np.repeat(self.full_fuel_consumed, self.variant_segments + 1, axis=0)
+        self.full_dist_traveled = np.repeat(self.full_dist_traveled, self.variant_segments + 1, axis=0)
+        self.time = np.repeat(self.time, self.variant_segments + 1, axis=0)
+        self.check_variant_def()
+
+        # determine new headings - centered around gcrs X0 -> X_prev_step
+        delta_hdgs = np.linspace(
+            -self.variant_segments/2 * self.variant_increments_deg,
+            +self.variant_segments/2 * self.variant_increments_deg,
+            self.variant_segments + 1)
+        delta_hdgs = np.tile(delta_hdgs, nof_input_routes)
+
+        self.current_variant = new_azi['azi1']	# center courses around gcr
+        self.current_variant = np.repeat(self.current_variant, self.variant_segments + 1)
+        self.current_variant = self.current_variant - delta_hdgs
+
+    def define_initial_variants(self):
+        pass
+
+    def move_boat_direct(self, wt : WeatherCond, boat: Boat, constraint_list: ConstraintsList):
+        """
+                calculate new boat position for current time step based on wind and boat function
+            """
+
+        # get wind speed (tws) and angle (twa)
+        debug = False
+
+        winds = self.get_wind_functions(wt) #wind is always a function of the variants
+        twa = winds['twa']
+        tws = winds['tws']
+        wind = {'tws': tws, 'twa': twa - self.get_current_azimuth()}
+        #if(debug) : print('wind in move_boat_direct', wind)
+
+        # get boat speed
+        bs = boat.boat_speed_function(wind)
+
+        ship_params = boat.get_fuel_per_time_netCDF(self.get_current_azimuth(), self.get_current_lats(),
+                                                  self.get_current_lons(), self.time, wind)
+        #ship_params.print()
+
+        delta_time, delta_fuel, dist = self.get_delta_variables_netCDF(ship_params, bs)
+        if (debug):
+            print('delta_time: ', delta_time)
+            print('delta_fuel: ', delta_fuel)
+            print('dist: ', dist)
+            print('is_last_step:', self.is_last_step)
+
+        move = self.check_bearing(dist)
+
+        if (debug):
+            print('move:', move)
+
+        if(self.is_last_step or self.is_pos_constraint_step):
+            delta_time, delta_fuel, dist = self.get_delta_variables_netCDF_last_step(ship_params, bs)
+
+        is_constrained = self.check_constraints(move, constraint_list)
+
+        self.update_position(move, is_constrained, dist)
+        self.update_time(delta_time)
+        self.update_fuel(delta_fuel)
+        self.update_shipparams(ship_params)
+        self.count += 1
+
+    def update_shipparams(self, ship_params_single_step):
+        new_rpm=np.vstack((ship_params_single_step.get_rpm(), self.shipparams_per_step.get_rpm()))
+        new_power=np.vstack((ship_params_single_step.get_power(), self.shipparams_per_step.get_power()))
+        new_speed=np.vstack((ship_params_single_step.get_speed(), self.shipparams_per_step.get_speed()))
+
+        self.shipparams_per_step.set_rpm(new_rpm)
+        self.shipparams_per_step.set_power(new_power)
+        self.shipparams_per_step.set_speed(new_speed)
+
 
     def check_variant_def(self):
         if (not ((self.lats_per_step.shape[1] == self.lons_per_step.shape[1]) and
@@ -104,7 +289,7 @@ class IsoBased(RoutingAlg):
             raise Exception('Pruned indices running out of bounds.')
 
     def pruning_per_step(self, trim = True):
-        # self.pruning_headings_centered(trim)
+        #self.pruning_headings_centered(trim)
         self.pruning_gcr_centered(trim)
 
     def pruning_gcr_centered(self, trim = True):
@@ -434,3 +619,27 @@ class IsoBased(RoutingAlg):
         self.starttime_per_step = np.expand_dims(self.starttime_per_step, axis=1)
 
         self.shipparams_per_step.expand_axis_for_intermediate()
+
+    def check_variant_def(self):
+        pass
+
+    def define_variants_per_step(self):
+        pass
+
+    def pruning_per_step(self, trim=True):
+        pass
+
+    def final_pruning(self):
+        pass
+
+    def get_current_azimuth(self):
+        pass
+
+    def update_dist(self, delta_time, bs):
+        pass
+
+    def update_time(self, delta_time, bs, current_lats, current_lons):
+        pass
+
+    def get_final_index(self):
+        pass
